refactor: merge DefaultParametersResourceMetadataCollectionFactory into ParameterValidationResourceMetadataCollectionFactory
