+++ src/Validator/Metadata/Resource/Factory/ParameterValidationResourceMetadataCollectionFactory.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Validator\Metadata\Resource\Factory;
 
+use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Parameter;
 use ApiPlatform\Metadata\Parameters;
@@ -30,6 +31,7 @@ final class ParameterValidationResourceMetadataCollectionFactory implements Reso
     public function __construct(
         private readonly ?ResourceMetadataCollectionFactoryInterface $decorated = null,
         private readonly ?ContainerInterface $filterLocator = null,
+        private readonly array $defaultParameters = [],
     ) {
     }
 
@@ -37,7 +39,11 @@ final class ParameterValidationResourceMetadataCollectionFactory implements Reso
     {
         $resourceMetadataCollection = $this->decorated?->create($resourceClass) ?? new ResourceMetadataCollection($resourceClass);
 
+        $defaultParams = $this->buildDefaultParameters();
+
         foreach ($resourceMetadataCollection as $i => $resource) {
+            $resource = $this->applyDefaults($resource, $defaultParams);
+
             $operations = $resource->getOperations();
 
             foreach ($operations as $operationName => $operation) {
@@ -134,5 +140,117 @@ final class ParameterValidationResourceMetadataCollectionFactory implements Reso
         }
 
         return $parameters;
+    }
+
+    /**
+     * Builds Parameter objects from the default configuration array.
+     *
+     * @return array<string, Parameter> Array of Parameter objects indexed by their key
+     */
+    private function buildDefaultParameters(): array
+    {
+        $parameters = [];
+
+        foreach ($this->defaultParameters as $parameterClass => $config) {
+            if (!is_subclass_of($parameterClass, Parameter::class)) {
+                continue;
+            }
+
+            $identifier = $config['key'] ?? (new \ReflectionClass($parameterClass))->getShortName();
+
+            $parameter = $this->createParameterFromConfig($parameterClass, $config);
+            $parameters[$identifier] = $parameter;
+        }
+
+        return $parameters;
+    }
+
+    /**
+     * Creates a Parameter instance from configuration.
+     *
+     * @param class-string<Parameter> $parameterClass The parameter class name
+     * @param array<string, mixed>    $config         The configuration array
+     *
+     * @return Parameter The created parameter instance
+     */
+    private function createParameterFromConfig(string $parameterClass, array $config): Parameter
+    {
+        return new $parameterClass(
+            key: $config['key'] ?? null,
+            schema: $config['schema'] ?? null,
+            openApi: null,
+            provider: null,
+            filter: $config['filter'] ?? null,
+            property: $config['property'] ?? null,
+            description: $config['description'] ?? null,
+            properties: null,
+            required: $config['required'] ?? false,
+            priority: $config['priority'] ?? null,
+            hydra: $config['hydra'] ?? null,
+            constraints: $config['constraints'] ?? null,
+            security: $config['security'] ?? null,
+            securityMessage: $config['security_message'] ?? null,
+            extraProperties: $config['extra_properties'] ?? [],
+            filterContext: null,
+            nativeType: null,
+            castToArray: null,
+            castToNativeType: null,
+            castFn: null,
+            default: $config['default'] ?? null,
+            filterClass: $config['filter_class'] ?? null,
+        );
+    }
+
+    /**
+     * Applies default parameters to the resource.
+     *
+     * @param array<string, Parameter> $defaultParams The default parameters to apply
+     */
+    private function applyDefaults(ApiResource $resource, array $defaultParams): ApiResource
+    {
+        $resourceParameters = $resource->getParameters() ?? new Parameters();
+        $mergedResourceParameters = $this->mergeParameters($resourceParameters, $defaultParams);
+        $resource = $resource->withParameters($mergedResourceParameters);
+
+        foreach ($operations = $resource->getOperations() ?? [] as $operationName => $operation) {
+            $operationParameters = $operation->getParameters() ?? new Parameters();
+            $mergedOperationParameters = $this->mergeParameters($operationParameters, $defaultParams);
+            $operations->add((string) $operationName, $operation->withParameters($mergedOperationParameters));
+        }
+
+        if ($operations) {
+            $resource = $resource->withOperations($operations);
+        }
+
+        foreach ($graphQlOperations = $resource->getGraphQlOperations() ?? [] as $operationName => $operation) {
+            $operationParameters = $operation->getParameters() ?? new Parameters();
+            $mergedOperationParameters = $this->mergeParameters($operationParameters, $defaultParams);
+            $graphQlOperations[$operationName] = $operation->withParameters($mergedOperationParameters);
+        }
+
+        if ($graphQlOperations) {
+            $resource = $resource->withGraphQlOperations($graphQlOperations);
+        }
+
+        return $resource;
+    }
+
+    /**
+     * Merges default parameters with operation-specific parameters.
+     *
+     * @param Parameters               $operationParameters The parameters already defined on the operation
+     * @param array<string, Parameter> $defaultParams       The default parameters to merge
+     *
+     * @return Parameters The merged parameters
+     */
+    private function mergeParameters(Parameters $operationParameters, array $defaultParams): Parameters
+    {
+        $merged = new Parameters($defaultParams);
+
+        foreach ($operationParameters as $key => $param) {
+            $merged->add($key, $param);
+        }
+
+        return $merged;
     }
 }
